Remove unused calendar import
