+++ backend/app/routes/sharing/__init__.py
@@ -1,0 +1,13 @@
+from . import invitations, notifications, shared_users, shared_users_medicines, grouped
+from .invitations import *
+from .notifications import *
+from .shared_users import *
+from .shared_users_medicines import *
+from .grouped import *
+
+__all__ = []
+__all__ += invitations.__all__ if hasattr(invitations, '__all__') else []
+__all__ += notifications.__all__ if hasattr(notifications, '__all__') else []
+__all__ += shared_users.__all__ if hasattr(shared_users, '__all__') else []
+__all__ += shared_users_medicines.__all__ if hasattr(shared_users_medicines, '__all__') else []
+__all__ += grouped.__all__ if hasattr(grouped, '__all__') else []
