+++ backend/app/routes/sharing/shared_users.py
@@ -1,7 +1,7 @@
 from app.utils.auth import require_auth
 from datetime import datetime, timezone
-from . import api
-from app.services.calendar import verify_calendar_share, verify_calendar, generate_calendar_schedule
+from .. import api
+from app.services.calendar import verify_calendar_share, generate_calendar_schedule
 from flask import request, g
 import time
 from app.services.user import fetch_user
@@ -12,7 +12,6 @@ import json
 from app.config import Config
 from urllib.parse import urljoin
 from app.services.medication import check_if_stock_is_low
-from app.services.notifications import email_address_direct
 
 
 ERROR_CALENDAR_NOT_FOUND = "calendrier non trouvé"
@@ -401,253 +400,3 @@ def handle_delete_user_shared_user(calendar_id, receiver_uid):
         )
 
 
-# fonction pour supprimer une invitation de calendrier partagé pour un user sans compte
-@api.route("/invitations/<calendar_id>", methods=["DELETE"])
-@require_auth
-def delete_shared_calendar_invitation(calendar_id):
-    try:
-        t_0 = time.time()
-        owner_uid = g.uid
-
-        if not verify_calendar(calendar_id, owner_uid):
-            return warning_response(
-                message=ERROR_UNAUTHORIZED_ACCESS,
-                code="UNAUTHORIZED_ACCESS",
-                status_code=404,
-                uid=owner_uid,
-                origin="GET_MEDICINE_BOXES",
-                log_extra={"calendar_id": calendar_id}
-            )
-
-        token = request.get_json(force=True).get("token")
-        receiver_email = request.get_json(force=True).get("email")
-
-        if not token:
-            return error_response(
-                message="Token de l'utilisateur requis",
-                code="MISSING_TOKEN",
-                status_code=400,
-                uid=g.uid,
-                origin="DELETE_SHARED_CALENDAR_INVITATION"
-            )
-        
-
-
-        with get_connection() as conn:
-            with conn.cursor() as cursor:
-                cursor.execute("DELETE FROM invitations WHERE token = %s AND calendar_id = %s", (token, calendar_id))
-                cursor.connection.commit()
-
-                email_address_direct(
-                    to_email=receiver_email,
-                    notification_type="calendar_invitation_registration_deleted",
-                    context={
-                        "sender_uid": owner_uid,
-                        "calendar_id": calendar_id,
-                    }
-                )
-                t_1 = time.time()
-
-                return success_response(
-                    message="Invitation de calendrier supprimée",
-                    code="SHARED_CALENDAR_INVITATION_DELETE_SUCCESS",
-                    uid=receiver_email,
-                    origin="DELETE_SHARED_CALENDAR_INVITATION",
-                    log_extra={"calendar_id": calendar_id, "time": t_1 - t_0}
-                )
-    except Exception as e:
-        return error_response(
-            message="Erreur lors de la récupération des données partagées",
-            code="SHARED_GROUPED_LOAD_ERROR",
-            status_code=500,
-            uid=owner_uid,
-            origin="GET_SHARED_GROUPED",
-            error=str(e)
-        )
-
-
-@api.route("/shared/grouped", methods=["GET"])
-@require_auth
-def handle_grouped_shared():
-    try:
-        t_0 = time.time()
-        uid = g.uid
-
-        with get_connection() as conn:
-            with conn.cursor() as cursor:
-                # Étape 1 : récupérer tous les calendriers personnels
-                cursor.execute("SELECT id, name FROM calendars WHERE owner_uid = %s", (uid,))
-                calendars = cursor.fetchall()
-                calendar_ids = [cal["id"] for cal in calendars]
-
-                grouped = {
-                    cal["id"]: {
-                        "calendar_name": cal["name"],
-                        "users": [],
-                        "invitation": [],
-                        "tokens": []
-                    }
-                    for cal in calendars
-                }
-
-                # Étape 2 : récupérer les utilisateurs partagés
-                cursor.execute("SELECT * FROM shared_calendars WHERE calendar_id = ANY(%s::uuid[])", (calendar_ids,))
-                shared_users = cursor.fetchall()
-
-                for shared in shared_users:
-                    calendar_id = shared["calendar_id"]
-                    receiver_uid = shared["receiver_uid"]
-
-                    if not verify_calendar_share(calendar_id, receiver_uid):
-                        continue
-
-                    receiver = fetch_user(receiver_uid)
-                    if not receiver:
-                        continue
-
-                    user_info = {
-                        "receiver_uid": receiver_uid,
-                        "access": shared.get("access", "read"),
-                        "accepted": shared.get("accepted", False),
-                        "receiver_photo_url": receiver.get("photo_url") or "https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/icons/person-circle.svg",
-                        "receiver_name": receiver.get("display_name"),
-                        "receiver_email": receiver.get("email")
-                    }
-
-                    if calendar_id in grouped:
-                        grouped[calendar_id]["users"].append(user_info)
-
-                # Étape 3 : récupérer les tokens partagés
-                cursor.execute("SELECT * FROM shared_tokens WHERE calendar_id = ANY(%s::uuid[])", (calendar_ids,))
-                tokens = cursor.fetchall()
-
-                for token in tokens:
-                    cal_id = token["calendar_id"]
-                    if cal_id in grouped:
-                        grouped[cal_id]["tokens"].append(token)
-
-                cursor.execute("SELECT * FROM invitations WHERE calendar_id = ANY(%s::uuid[])", (calendar_ids,))
-                invitations = cursor.fetchall()
-
-                for invitation in invitations:
-                    cal_id = invitation["calendar_id"]
-                    if cal_id in grouped:
-                        grouped[cal_id]["invitation"].append(invitation)
-
-        t_1 = time.time()
-        return success_response(
-            message="Données partagées groupées récupérées",
-            code="SHARED_GROUPED_LOAD_SUCCESS",
-            uid=uid,
-            origin="SHARED_GROUPED_LOAD",
-            data={"grouped": grouped},
-            log_extra={"calendar_count": len(grouped), "time": t_1 - t_0}
-        )
-
-    except Exception as e:
-        return error_response(
-            message="Erreur lors du groupement des données partagées",
-            code="SHARED_GROUPED_LOAD_ERROR",
-            status_code=500,
-            uid=g.uid,
-            origin="SHARED_GROUPED_LOAD",
-            error=str(e)
-        )
-
-
-@api.route("/shared/users/calendars/<calendar_id>/notifications", methods=["GET"])
-@require_auth
-def handle_shared_user_notifications(calendar_id):
-    try:
-        t_0 = time.time()
-        uid = g.uid
-        if not verify_calendar_share(calendar_id, uid):
-            return warning_response(
-                message=ERROR_UNAUTHORIZED_ACCESS,
-                code="SHARED_CALENDARS_NOTIFICATIONS_ERROR",
-                status_code=403,
-                uid=uid,
-                origin="SHARED_CALENDARS_NOTIFICATIONS",
-                log_extra={"calendar_id": calendar_id}
-            )
-        with get_connection() as conn:
-            with conn.cursor() as cursor:
-                cursor.execute("SELECT notifications_enabled FROM shared_calendars WHERE calendar_id = %s", (calendar_id,))
-                calendar = cursor.fetchone()
-                if not calendar:
-                    return warning_response(
-                        message="calendrier partagé non trouvé",
-                        code="SHARED_CALENDARS_NOTIFICATIONS_ERROR",
-                        status_code=404,
-                        uid=uid,
-                        origin="SHARED_CALENDARS_NOTIFICATIONS",
-                        log_extra={"calendar_id": calendar_id}
-                    )
-                notifications_enabled = calendar.get("notifications_enabled", False)
-        t_1 = time.time()
-        return success_response(
-            message="notifications récupérées",
-            code="SHARED_CALENDARS_NOTIFICATIONS_SUCCESS",
-            uid=uid,
-            origin="SHARED_CALENDARS_NOTIFICATIONS",
-            data={"notifications-enabled": notifications_enabled},
-            log_extra={"calendar_id": calendar_id, "time": t_1 - t_0}
-        )
-    except Exception as e:
-        return error_response(
-            message="erreur lors de la récupération des notifications",
-            code="SHARED_CALENDARS_NOTIFICATIONS_ERROR",
-            status_code=500,
-            uid=uid,
-            origin="SHARED_CALENDARS_NOTIFICATIONS",
-            error=str(e),
-            log_extra={"calendar_id": calendar_id}
-        )
-
-@api.route("/shared/users/calendars/<calendar_id>/notifications", methods=["PUT"])
-@require_auth
-def handle_shared_user_notifications_update(calendar_id):
-    try:
-        t_0 = time.time()
-        uid = g.uid
-        if not verify_calendar_share(calendar_id, uid):
-            return warning_response(
-                message=ERROR_UNAUTHORIZED_ACCESS,
-                code="SHARED_CALENDARS_NOTIFICATIONS_ERROR",
-                status_code=403,
-                uid=uid,
-                origin="SHARED_CALENDARS_NOTIFICATIONS",
-                log_extra={"calendar_id": calendar_id}
-            )
-        data = request.get_json()
-        if not data or "notifications-enabled" not in data:
-            return warning_response(
-                message="Données manquantes",
-                code="SHARED_CALENDARS_NOTIFICATIONS_ERROR",
-                status_code=400,
-                uid=uid,
-                origin="SHARED_CALENDARS_NOTIFICATIONS",
-                log_extra={"calendar_id": calendar_id}
-            )
-        notifications_enabled = data["notifications-enabled"]
-        with get_connection() as conn:
-            with conn.cursor() as cursor:
-                cursor.execute("UPDATE shared_calendars SET notifications_enabled = %s WHERE calendar_id = %s", (notifications_enabled, calendar_id))
-        t_1 = time.time()
-        return success_response(
-            message="Notifications mises à jour",
-            code="SHARED_CALENDARS_NOTIFICATIONS_SUCCESS",
-            uid=uid,
-            origin="SHARED_CALENDARS_NOTIFICATIONS",
-            log_extra={"calendar_id": calendar_id, "time": t_1 - t_0}
-        )
-    except Exception as e:
-        return error_response(
-            message="Erreur lors de la mise à jour des notifications",
-            code="SHARED_CALENDARS_NOTIFICATIONS_ERROR",
-            status_code=500,
-            uid=uid,
-            origin="SHARED_CALENDARS_NOTIFICATIONS",
-            error=str(e),
-            log_extra={"calendar_id": calendar_id}
-        )
